Squashed commit of the following:

commit 02d7526cd04fa76e2406462c3a6c376b28cef431
Merge: 5e2620b48 f644f06c3
Author: Murad Mamedov <murad.mamedov@relativity.com>
Date:   Thu Nov 28 11:33:26 2019 +0100

    Merge remote-tracking branch 'origin/develop' into REL-378363-could-not-convert-string-to-integer-2222222222222222222

commit 5e2620b485d7d898a67eda21170dd1638a8870b4
Author: Murad Mamedov <murad.mamedov@relativity.com>
Date:   Mon Nov 25 16:00:05 2019 +0100

    Added validation to UI and changed type of VolumeMaxSize from int to long.

+++ Source/kCura.IntegrationPoints.Core/Models/ExportSettings.cs
@@ -95,7 +95,7 @@ namespace kCura.IntegrationPoints.FilesDestinationProvider.Core
 		public string VolumePrefix { get; set; }
 		public int VolumeStartNumber { get; set; }
 		public int VolumeDigitPadding { get; set; }
-		public int VolumeMaxSize { get; set; }
+		public long VolumeMaxSize { get; set; }
 		public FilePathType FilePath { get; set; }
 		public string UserPrefix { get; set; }
 		public bool ExportMultipleChoiceFieldsAsNested { get; set; }
